requirements.txt
pytest
pytest-mock
bcrypt
pyjwt
pydantic[email]
pydantic
mysql-connector-python
flask 
python-dotenv
redis